drf/3.2-crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter

